Use Control.Exception.evaluate instead of `pure $!`
@sol points out, in
https://github.com/benclifford/acme-smuggler/commit/ec47172b6d450077e21ea4f2fda192bda2c6acf9,
that evaluate has different exception semantics to `pure $!`
as documented in the Control.Exception.evaluate haddocks.

However, I think (without thinking about it too much) that that
difference doesn't matter in this particular circumstance as we
know everything that is being evaluated within the `try` block.

Using `evaluate` is a bit easier on the eye too.